UPD(dtests): Updated files for use on Windows
~~
Cause:
  We are NOT able to test Reactive C++ client on Windows.
Consequence:
  Reactive C++ client on Windows jobs does not run in CI etc.
Fix:
  CMakeLists.txt and some source codes was updated to run on Windows.
Result:
  We are able to test Reactive C++ client on Windows.
~~
Relates:
  MSGQE-3269 [reactive-cpp] run reactive cpp tests (333) on Windows in CI
  MSGQE-4375 [reactive-cpp] help Radim run reactive cpp tests (333) on Windows in CI
GitOrigin-RevId: f3808e2cd1f2f9e952d81f115c7298457445ceb9

+++ dtests/node_data/clients/cpp/src/api/qpid-proton/reactor/formatter/ReactorDecoder.cpp
@@ -248,7 +248,7 @@ void ReactorDecoder::decodeProperties(Writer *writer) const
     
     message::property_map props = m.properties();
 
-    logger(debug) << "Number of properties: " << props.size();
+    logger(debug) << "Number of properties: " << (unsigned long) props.size();
 
     writer->startProperties();
     writeProperties(writer);
@@ -369,7 +369,7 @@ string ReactorDecoder::decodeValue(const value &value) const {
             value.get(l);
             logger(debug) << "(m) Type id: " << type;
 
-            logger(debug) << "(m) Size: " << l.size();
+            logger(debug) << "(m) Size: " << (unsigned long) l.size();
 
             for (size_t i = 0; i < l.size(); i++) {
                 // decodeValue(writer, dec);
